organised code into the src directory
