+++ src/models/Nptg.ts
