+++ test/custom_tests/testLocationNames.js
